cleanup/refactor

+++ api-admin/src/main/java/bio/terra/pearl/api/admin/service/enrollee/EnrolleeImportExtService.java
@@ -4,15 +4,18 @@ import bio.terra.pearl.api.admin.service.AuthUtilService;
 import bio.terra.pearl.core.model.EnvironmentName;
 import bio.terra.pearl.core.model.admin.AdminUser;
 import bio.terra.pearl.core.model.dataimport.Import;
+import bio.terra.pearl.core.model.study.PortalStudy;
 import bio.terra.pearl.core.model.study.StudyEnvironment;
 import bio.terra.pearl.core.service.CascadeProperty;
 import bio.terra.pearl.core.service.dataimport.ImportService;
+import bio.terra.pearl.core.service.exception.PermissionDeniedException;
 import bio.terra.pearl.core.service.export.EnrolleeImportService;
 import bio.terra.pearl.core.service.study.StudyEnvironmentService;
+import org.springframework.stereotype.Service;
+
 import java.io.InputStream;
 import java.util.List;
 import java.util.UUID;
-import org.springframework.stereotype.Service;
 
 @Service
 public class EnrolleeImportExtService {
@@ -40,19 +43,26 @@ public class EnrolleeImportExtService {
   public Import importData(
       String portalShortcode,
       String studyShortcode,
-      String envName,
+      EnvironmentName environmentName,
       InputStream tsvData,
       AdminUser operator) {
     authUtilService.authUserToPortal(operator, portalShortcode);
-    EnvironmentName environmentName = EnvironmentName.valueOfCaseInsensitive(envName);
-    StudyEnvironment studyEnv =
-        studyEnvironmentService.verifyStudy(studyShortcode, environmentName);
+    PortalStudy portalStudy = authUtilService.authUserToStudy(operator, portalShortcode, studyShortcode);
+    StudyEnvironment studyEnv = studyEnvironmentService.verifyStudy(studyShortcode, environmentName);
     return enrolleImportService.importEnrollees(
         portalShortcode, studyShortcode, studyEnv, tsvData, operator.getId());
   }
 
-  public void delete(String portalShortcode, UUID id, AdminUser operator) {
+  public void delete(String portalShortcode, String studyShortcode,
+                     EnvironmentName environmentName, UUID id, AdminUser operator) {
     authUtilService.authUserToPortal(operator, portalShortcode);
+    PortalStudy portalStudy = authUtilService.authUserToStudy(operator, portalShortcode, studyShortcode);
+    StudyEnvironment studyEnv = studyEnvironmentService.verifyStudy(studyShortcode, environmentName);
+    Import dataImport = importService.find(id).get();
+    if (!dataImport.getStudyEnvironmentId().equals(studyEnv.getId())) {
+      throw new PermissionDeniedException("Import Id does not belong to the given study environment");
+    }
     importService.delete(id, CascadeProperty.EMPTY_SET);
   }
+
 }
